donizo-material-scraper/requirements.txt
requests
beautifulsoup4
selenium
PyYAML
pytest
fastapi
uvicorn
playwright 
python-dotenv
pytest
streamlit